reverted portibility fix, tried new one

+++ analysis/include/UnorderedJetTagFilter.hh
@@ -1,6 +1,7 @@
 #ifndef UNORDERED_JET_TAG_FILTER_HH
 #define UNORDERED_JET_TAG_FILTER_HH
 
+#include <cstddef>
 #include <vector> 
 #include <map>
 #include "btag_enums.hh"
@@ -19,8 +20,8 @@ public:
   virtual double jet_scalefactor(const std::vector<Jet>& jets) const; 
 private: 
   syst::Systematic m_systematic; 
-  unsigned m_n_jets_skipped; 
-  std::map<btag::OperatingPoint,unsigned> m_tag_counts; 
+  size_t m_n_jets_skipped; 
+  std::map<btag::OperatingPoint,size_t> m_tag_counts; 
 };
 
 namespace jetfilter {
